xserver xkeyboard-config libX11 fontconfig mesa putty git update Oct 10 2021
xserver          commit ca1dfdc9aa4b548de624d3a9af5147a998ba3d79
xkeyboard-config commit 8a03e98d5b256ac882db2c0ed88dcf5180c87a1d
libX11           commit 2356e59ff24f8d1b25cdc4dffc5171c65dc2b86e
fontconfig       commit 1d7669992798bb7fec7dc613c05310741d7b3781
mesa             commit 1c0eb7aa78bac4494dbff0c502a8c09e23a4c123
putty            commit 54930cf784163294ce1211b9d7c3c91b372e13ef

+++ mesalib/src/gallium/drivers/r300/r300_state.c
@@ -1821,7 +1821,7 @@ static void r300_vertex_psc(struct r300_vertex_element_state *velems)
 
         if (i & 1) {
             vstream->vap_prog_stream_cntl[i >> 1] |= type << 16;
-            vstream->vap_prog_stream_cntl_ext[i >> 1] |= swizzle << 16;
+            vstream->vap_prog_stream_cntl_ext[i >> 1] |= (uint32_t)swizzle << 16;
         } else {
             vstream->vap_prog_stream_cntl[i >> 1] |= type;
             vstream->vap_prog_stream_cntl_ext[i >> 1] |= swizzle;
